requirements.txt
mysql-connector-python>=8.0.0
Django>=3.2.0
python-dotenv>=0.19.0